moved the plotting part of simulacovid to plots.py and moved the time conversion function to utils
